scripts/compute_mappings/tests/requirements.txt
pytest-subtests